+++ gradient/api_sdk/repositories/tensorboards.py
@@ -1,11 +1,11 @@
-from gradient.api_sdk.config import config
 from . import common
+from .. import config
 from .. import serializers, models
 
 
 class GetTensorboardApiUrlMixin(object):
     def _get_api_url(self, **kwargs):
-        return config.CONFIG_SERVICE_HOST
+        return config.config.CONFIG_SERVICE_HOST
 
 
 class ParseTensorboardMixin(object):
