+++ gradient/api_sdk/serializers/deployment.py
@@ -1,7 +1,7 @@
 import marshmallow as ma
 
-from gradient.api_sdk import models
 from .base import BaseSchema
+from .. import models
 
 
 class DeploymentSchema(BaseSchema):
