+++ gradient/api_sdk/utils.py
@@ -1,4 +1,5 @@
 import base64
+import os
 import random
 import string
 from collections import OrderedDict
@@ -141,3 +142,42 @@ class MultipartEncoderWithProgressbar(MultipartEncoder):
                 bar.update(monitor.bytes_read)
 
         return callback
+
+
+class PathParser(object):
+    LOCAL_DIR = 0
+    LOCAL_FILE = 1
+    GIT_URL = 2
+    S3_URL = 3
+
+    @classmethod
+    def parse_path(cls, path):
+        if cls.is_local_dir(path):
+            return cls.LOCAL_DIR
+
+        if cls.is_local_zip_file(path):
+            return cls.LOCAL_FILE
+
+        if cls.is_git_url(path):
+            return cls.GIT_URL
+
+        if cls.is_s3_url(path):
+            return cls.S3_URL
+
+        raise sdk_exceptions.WrongPathError("Given path is neither local path, nor valid URL")
+
+    @staticmethod
+    def is_local_dir(path):
+        return os.path.exists(path) and os.path.isdir(path)
+
+    @staticmethod
+    def is_local_zip_file(path):
+        return os.path.exists(path) and os.path.isfile(path) and path.endswith(".zip")
+
+    @staticmethod
+    def is_git_url(path):
+        return not os.path.exists(path) and path.endswith(".git") or path.lower().startswith("git:")
+
+    @staticmethod
+    def is_s3_url(path):
+        return not os.path.exists(path) and path.lower().startswith("s3:")
