+++ gradient/cliutils.py
@@ -1,12 +1,10 @@
 import json
-import os
 import shutil
 
 import requests
 import six
 
 from gradient import exceptions
-from gradient.exceptions import WrongPathError
 
 
 def get_terminal_lines(fallback=48):
@@ -37,45 +35,6 @@ def status_code_to_error_obj(status_code):
     if status_code in requests.status_codes._codes:
         message = requests.status_codes._codes[status_code][0]
     return {'error': True, 'message': message, 'status': status_code}
-
-
-class PathParser(object):
-    LOCAL_DIR = 0
-    LOCAL_FILE = 1
-    GIT_URL = 2
-    S3_URL = 3
-
-    @classmethod
-    def parse_path(cls, path):
-        if cls.is_local_dir(path):
-            return cls.LOCAL_DIR
-
-        if cls.is_local_zip_file(path):
-            return cls.LOCAL_FILE
-
-        if cls.is_git_url(path):
-            return cls.GIT_URL
-
-        if cls.is_s3_url(path):
-            return cls.S3_URL
-
-        raise WrongPathError("Given path is neither local path, nor valid URL")
-
-    @staticmethod
-    def is_local_dir(path):
-        return os.path.exists(path) and os.path.isdir(path)
-
-    @staticmethod
-    def is_local_zip_file(path):
-        return os.path.exists(path) and os.path.isfile(path) and path.endswith(".zip")
-
-    @staticmethod
-    def is_git_url(path):
-        return not os.path.exists(path) and path.endswith(".git") or path.lower().startswith("git:")
-
-    @staticmethod
-    def is_s3_url(path):
-        return not os.path.exists(path) and path.lower().startswith("s3:")
 
 
 def validate_auth_options(kwargs):
